ai_service/pyproject.toml
[tool.poetry]
name = "ai_service"
version = "0.1.0"
description = "This project is designed to analyze student achievements and identify potential risks."
authors = ["Viet Hong Bui <henrygphp@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
scikit-learn = "^1.6.1"
pandas = "^2.2.3"
aioodbc = "^0.5.0"
joblib = "^1.4.2"
python-dotenv = "^1.1.0"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
streamlit = "^1.44.0"
matplotlib = "^3.10.1"
httpx = "^0.28.1"
pyodbc = "^5.2.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
databases = "^0.9.0"
pydantic = "^2.11.1"
scalar-fastapi = "^1.0.3"
fastmcp = "^0.4.1"
nest-asyncio = "^1.6.0"
motor = "^3.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"